execution/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-api-python-client>=2.100.0

# RAG Pipeline (PRD 01)
sentence-transformers>=2.7.0
supabase>=2.0.0
numpy>=1.24.0
firecrawl-py>=1.2.0